Avoid arrow functions in render()
The arrow functions means that React has to re-render the component on
every cycle given that there's no way to check that the inline arrow
functions are constant between renders. As such, it's better to avoid
arrow functions in render() since they cause performance issues. There
are several ways to get around the arrow functions, refer to this
article for more context:
https://medium.freecodecamp.org/react-binding-patterns-5-approaches-for-handling-this-92c651b5af56

Here, we go for the "arrow functions in class property" approach which
is recommended by Facebook for non-legacy code:
https://facebook.github.io/react/docs/react-without-es6.html#autobinding